Made IEphemeralEvent implement IClientTimestamp, and moved isNewer logic back to EphemeralEvent
