+++ packages/live-share/src/EphemeralEvent.ts
@@ -6,7 +6,7 @@
 import { DataObject, DataObjectFactory } from '@fluidframework/aqueduct';
 import { IEvent } from "@fluidframework/common-definitions";
 import { LocalTimestampProvider } from "./LocalTimestampProvider";
-import { IEphemeralEvent, ITimestampProvider, IRoleVerifier, UserMeetingRole } from "./interfaces";
+import { IEphemeralEvent, ITimestampProvider, IRoleVerifier, UserMeetingRole, IClientTimestamp } from "./interfaces";
 import { EphemeralEventScope } from './EphemeralEventScope';
 import { EphemeralEventTarget } from './EphemeralEventTarget';
 import { LocalRoleVerifier } from './LocalRoleVerifier';
@@ -167,8 +167,32 @@ export interface IEphemeralEventEvents<TEvent extends IEphemeralEvent> extends I
      * @param debouncePeriod Optional. Time in milliseconds to ignore any new events for. Defaults to 0 ms.
      * @returns True if the received event is newer then the current event and should replace the current one. 
      */
-    public static isNewer(current: IEphemeralEvent|undefined, received: IEphemeralEvent, debouncePeriod = 0): boolean {
-        return isNewer(current, received, debouncePeriod)
+    public static isNewer(current: IClientTimestamp|undefined, received: IClientTimestamp, debouncePeriod = 0): boolean {
+        if (current) {
+            if (current.timestamp == received.timestamp) {
+                // In a case where both clientId's are blank that's the local client in a disconnected state
+                const cmp = (current.clientId || '').localeCompare(received.clientId || '');
+                if (cmp < 0) {
+                    // cmp == 0 is same user and we want to take latest event from a given user.
+                    // cmp > 0 is a tie breaker so we'll take that event as well (comparing 'a' with 'c' 
+                    // will result in a negative value).
+                    return false;
+                }
+            } else if (current.timestamp > received.timestamp) {
+                // Did we receive an older event that should have caused us to debounce the current one?
+                const delta = current.timestamp - received.timestamp;
+                if (delta > debouncePeriod) {
+                    return false;
+                }
+            } else {
+                // Is the new event within the debounce period?
+                const delta =  received.timestamp - current.timestamp;
+                if (delta < debouncePeriod) {
+                    return false;
+                }
+            }
+        }
+        return true
     }
 
     /**
